+++ R/collect_media.R
@@ -23,6 +23,7 @@ collect_media_metadata <- function(.query){
 #' Internal version of `collect()` for `request_files(type = "media")`
 #' @param object of class `files_response`, from `compute()`
 #' @importFrom dplyr group_by
+#' @importFrom dplyr filter
 #' @importFrom dplyr count
 #' @importFrom rlang .data
 #' @noRd
@@ -33,14 +34,30 @@ collect_media_files <- function(.query){
     status_code = unlist(lapply(result, function(a){a$status_code}))) |>
     group_by(.data$status_code) |>
     count()
+  
+  # successful downloads
   success <- result_summary |> 
-    dplyr::filter(result_summary$status_code == 200) 
+    dplyr::filter(.data$status_code == 200) 
   n_downloaded <- success[["n"]]
+  
+  # failed downloads
+  fail <- NULL
+  if(nrow(result_summary) > 1) {
+    if(any(result_summary$status_code %in% "403")) {
+      fail <- result_summary |>
+        dplyr::filter(.data$status_code == 403)
+      n_failed <- fail[["n"]]
+    }
+  }
   user_directory <- pour("package", "directory")
   bullets <- c(
     "v" = glue("Downloaded {n_downloaded} files successfully (status 200)."),
     ">" = glue("Files saved in local directory: \"{user_directory}\".")
   )
+  if(!is.null(fail)) {
+    bullets <- c(bullets,
+                 "x" = glue("Failed {n_failed} downloads due to missing images (status 403)"))
+  }
   inform(bullets)
   invisible(result_summary)
 }
@@ -93,7 +110,7 @@ collect_media <- function(df,
   user_directory <- pour("package", "directory")
   if (stringr::str_detect(user_directory, "Temp")) {
     inform(
-      cli::cli_text("{cli::col_magenta('To change which file directory media files are saved, use `galah_config(directory = )`.')}")
+      cli::cli_text("{cli::col_magenta('To change which file directory media files are saved to, use `galah_config(directory = )`.')}")
       )
   }