+++ R/parse_occurrences_count.R
@@ -9,7 +9,7 @@
 #' @param .query An object of class `data_query`
 #' @keywords Internal
 #' @noRd
-compute_occurrences_count <- function(.query){
+parse_occurrences_count <- function(.query){
   if(is_gbif()){
     if(.query$expand){    
       abort("Grouped counts haven't been (re)implemented for GBIF yet")
@@ -19,9 +19,9 @@ compute_occurrences_count <- function(.query){
     }
   }else{
     if(.query$expand){
-      compute_occurrences_count_groupby(.query)
+      parse_occurrences_count_groupby(.query)
     }else{
-      compute_occurrences_count_nogroupby(.query)
+      parse_occurrences_count_nogroupby(.query)
     }
   }
 }
@@ -31,18 +31,37 @@ compute_occurrences_count <- function(.query){
 #' Internal function to handle facet counting, adjustment etc.
 #' @noRd
 #' @keywords Internal
-compute_occurrences_count_nogroupby <- function(.query){
+parse_occurrences_count_nogroupby <- function(.query){
   url <- url_parse(.query$url)
+  
+  # check if a limit has been set
   if(!is.null(url$query$flimit)){
+    
+    # check number of total facets in the field
+    n_facets <- check_facet_count(.query)
+    
+    # handle slice_head
     if(as.integer(url$query$flimit) < 1){
-      n_facets <- check_facet_count(.query)
       url$query$flimit <- .query$arrange$slice_n # Q: is this correct? 
       if(.query$arrange$slice_n < n_facets){
         url$query$foffset <- n_facets - .query$arrange$slice_n
       }
       .query$url <- url_build(url) 
       .query
+    
+    # message when limit is hit
     }else{
+      if(as.integer(url$query$flimit) < n_facets){
+        limit <- url$query$flimit |> prettyNum(big.mark=",", preserve.width="none")
+        n_total_facets <- n_facets |> prettyNum(big.mark=",", preserve.width="none")
+        
+        bullets <- c(
+          cli::cli_text(cli::col_yellow(glue("Limiting to first {limit} of {n_total_facets} rows."))),
+          cli::cli_text(cli::col_magenta("Use `atlas_counts(limit = )` to return more rows."))
+        )
+        inform(bullets)
+      }
+      # .query$url <- url_build(url) 
       .query
     }
   }else{
@@ -62,7 +81,7 @@ compute_occurrences_count_nogroupby <- function(.query){
 #' @importFrom httr2 url_parse
 #' @noRd
 #' @keywords Internal
-compute_occurrences_count_groupby <- function(.query, error_call = caller_env()){
+parse_occurrences_count_groupby <- function(.query, error_call = caller_env()){
   data_cached <- .query
   # get url
   url <- url_parse(.query$url)
@@ -91,7 +110,7 @@ compute_occurrences_count_groupby <- function(.query, error_call = caller_env())
   saved_facet_queries <- saved_facet_queries[
     unlist(lapply(saved_facet_queries, function(a){!is.null(a)}))]
   
-  # rebuild url
+  # rebuild url with only first facet argument
   url$query <- c(
     url$query[names(url$query) != "facets"],
     facet_list[-length(facet_list)])