+++ tests/testthat/test-international-Spain.R
@@ -7,7 +7,9 @@ test_that("swapping to atlas = Spain works", {
 
 test_that("show_all(fields) works for Spain", {
   skip_if_offline()
-  x <- show_all(fields)
+  x <- show_all(fields) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gt(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
   # also that fields match those returned by default_columns()
@@ -17,70 +19,90 @@ test_that("show_all(fields) works for Spain", {
 
 test_that("show_all(licences) works for Spain", {
   skip_if_offline()
-  x <- show_all(licences)
+  x <- show_all(licences) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gt(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("show_all(collections) works for Spain", {
   skip_if_offline()
-  x <- show_all(collections, limit = 10)
+  x <- show_all(collections, limit = 10) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_lte(nrow(x), 10)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("show_all(datasets) works for Spain", {
   skip_if_offline()
-  x <- show_all(datasets, limit = 10)
+  x <- show_all(datasets, limit = 10) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_lte(nrow(x), 10)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("show_all(providers) works for Spain", {
   skip_if_offline()
-  x <- show_all(providers, limit = 10)
+  x <- show_all(providers, limit = 10) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_lte(nrow(x), 10)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("show_all(reasons) works for Spain", {
   skip_if_offline()
-  x <- show_all(reasons)
+  x <- show_all(reasons) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gt(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("show_all(assertions) works for Spain", {
   skip_if_offline()
-  x <- show_all(assertions)
+  x <- show_all(assertions) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gt(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("show_all(profiles) works for Spain", {
   skip_if_offline()
-  x <- show_all(profiles)
+  x <- show_all(profiles) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gt(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("show_all(lists) works for Spain", {
   skip_if_offline()
-  x <- show_all(lists, limit = 10)
+  x <- show_all(lists, limit = 10) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_lte(nrow(x), 10)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("search_all(fields) works for Spain", {
   skip_if_offline()
-  x <- search_all(fields, "year")
+  x <- search_all(fields, "year") |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gte(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("search_all(taxa) works for Spain", {
   skip_if_offline()
-  x <- search_all(taxa, "Mammalia")
+  x <- search_all(taxa, "Mammalia") |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gte(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
@@ -89,14 +111,18 @@ test_that("search_all(taxa) works using data.frames for Spain", {
   skip_if_offline()
   x <- search_all(taxa, 
                   data.frame(kingdom = "Animalia", 
-                             phylum = "Chordata"))
+                             phylum = "Chordata")) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gte(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("search_all(identifiers) works for Spain", {
   skip_if_offline()
-  x <- search_all(identifiers, "359")
+  x <- search_all(identifiers, "359") |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gte(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
@@ -104,7 +130,9 @@ test_that("search_all(identifiers) works for Spain", {
 test_that("show_values works for fields for Spain", {
   skip_if_offline()
   x <- search_all(fields, "basisOfRecord") |> 
-    show_values()
+    show_values() |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gte(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
@@ -112,15 +140,29 @@ test_that("show_values works for fields for Spain", {
 test_that("show_values works for profiles for Spain", {
   skip_if_offline()
   x <- search_all(profiles, "LA") |> 
-    show_values()
+    show_values() |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
   expect_gte(nrow(x), 1)
   expect_true(inherits(x, c("tbl_df", "tbl", "data.frame")))
 })
 
 test_that("atlas_counts works for Spain", {
   skip_if_offline()
-  expect_gt(atlas_counts()$count, 0)
-  expect_gt(atlas_counts(type = "species")$count, 0)
+  x <- atlas_counts() |>
+    pull(count) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
+  expect_gt(x, 0)
+})
+
+test_that("atlas_counts works with type = 'species' for Spain", {
+  skip_if_offline()
+  x <- atlas_counts(type = "species") |>
+    pull(count) |>
+    try(silent = TRUE)
+  skip_if(inherits(x, "try-error"), message = "API not available")
+  expect_gt(x, 0)
 })
 
 test_that("atlas_counts works with galah_identify for Spain", {
@@ -128,12 +170,16 @@ test_that("atlas_counts works with galah_identify for Spain", {
   result <- galah_call() |>
     identify("Mammalia") |>
     count() |>
-    collect()
+    collect() |>
+    try(silent = TRUE)
+  skip_if(inherits(result, "try-error"), message = "API not available")
   expect_gt(result$count, 1)
   result2 <- galah_call() |>
     filter(class == "Mammalia") |>
     count() |>
-    collect()
+    collect() |>
+    try(silent = TRUE)
+  skip_if(inherits(result2, "try-error"), message = "API not available")
   expect_lt(
     sqrt((result2$count - result$count)^2) / result$count, 
     0.1) # i.e. <1% margin of error
@@ -145,7 +191,9 @@ test_that("atlas_counts works with group_by for Spain", {
     filter(year >= 2000) |>
     group_by(basis_of_record) |>
     count() |>
-    collect()
+    collect() |>
+    try(silent = TRUE)
+  skip_if(inherits(result, "try-error"), message = "API not available")
   expect_gt(nrow(result), 1)
   expect_equal(names(result), c("basis_of_record", "count"))
   })
@@ -192,7 +240,9 @@ test_that("atlas_occurrences works for Spain", {
     galah_identify("Mammalia") |>
     galah_filter(year <= 1800) |>
     galah_select(species, year) |>
-    atlas_occurrences()
+    atlas_occurrences() |>
+    try(silent = TRUE)
+  skip_if(inherits(occ, "try-error"), message = "API not available")
   expect_gt(nrow(occ), 0)
   expect_equal(ncol(occ), 2)
   expect_true(inherits(occ, c("tbl_df", "tbl", "data.frame")))